Update DefaultConnection to use SQL Server Express instance
The connection string for the `DefaultConnection` in the `appsettings.json` file has been modified. The server name has been changed from `.` (which refers to the default SQL Server instance) to `.\SQLEXPRESS` (which refers to a named instance of SQL Server Express). This change ensures that the application connects to the SQL Server Express instance instead of the default SQL Server instance.